Janus webapp snapshot, preparing to publish the repo
This is a snapshot in time of the Janus webapp and the supporting
configTools library. This repo contains the source code with the
Guardian's data removed and no history.

Going forwards, we expect this to be the place where changes to the
webapp are made, while changes to the Guardian-specific data and
configuration can be made elsewhere.