Tidy up repository
Removes a couple more Guardian-specific mentions from the
tempplates. Crucially, this means that we need to configure a
reference to the source of Janus Data configuration. This allows the
UI to provide helpful feedback to users, without being hard-coded to a
repository. The config format is backwards compatible so no changes
are currently required in our Janus Data, and so that teams using
Janus are not required to configure this property.

The help template was particularly unhelpful, containing out-of date
information. The template (and its route) has been completely removed
from the project. We should think about adding something more accurate
in future.