Bump https://github.com/typelevel/kind-projector
This is required as part of the Scala 2.13.10 migration.

Note, this library results in horrible errors and should probably
be dropped. For now, I've fixed some errors from upgrading by moving
from ? to *.

The 'target' Scalac flag has also been switched to 'release' as the
former is now deprecated and was causing the build to fail.