ci: Tighten the build
This change removes the `janus-app-branch` input to the Workflow being triggered in `guardian/janus`.

The build is now:
  1. Compile the Scala app, and run the tests
  2. If compilation is successful, and we're on the `main` branch, trigger the Workflow in `guardian/janus`
  3. The Workflow in `guardian/janus` will clone always clone the main branch of this repository

With these changes we lose the ability to deploy feature branches (as their build will not appear in Riff-Raff),
however looking at the deployment history, this isn't a thing we've ever done.

Co-authored-by: Natasha <novembertang@users.noreply.github.com>
Co-authored-by: Adam <adamnfish@users.noreply.github.com>