Improve the application version comment
We pin the application version in the Debian scope, so that the app's
artifact (a .deb) has a name that matches the userdata in the
cloudformation. This documentation block is to help someone that comes
afterwards understand the relationship between the two versions in the
project, and the cloudformation template in another repository.