Increase the time we wait for the private build to complete
The real build gets kicked off in another repository, and then we have
a step that waits for a while before checking if it succeeded. In
practice, with a cold cache we observe that a 6 minute wait covers all
cases, so we've upped the wait time.

Note: Better would be to poll until the build completes, but since
this doesn't actually delay anything, no harm in having it be long
enough that it reliably gives the correct answer.