Refactor auto-signout logic
Pulls auto-signout redirect URL generation logic out for testing, and
so it can be reused across both console URL endpoints. We can generate
a consoel URL for the user, or redirect them directly. In both cases,
we should respect the user's auto-logout setting.

I've kept this as a cookie for now, we can always revisit whether it
would make sense to expose this in the URL in future. The key is that
I think it makes sense for existing URLs to respect the user's
behaviour, especially if they have been bookmarked.