feat: use html5 datepicker
Most browsers now offer a datepicker element by default (see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/date).

The original 'pickadate' function in Materialize used an external library which in v1 was replaced by a system-specific 'datepicker' function.
However, the options for this function differ from the original inasmuch as to reduce compatibility of the code. The naming of this
function is now the same as jQuery's datepicker, and this caused additional issues.
I have decided to go with the default HTML5 datepicker input element as it is
"well established and works across many devices and browser versions.
It’s been available across browsers since April 2021."
It was also simpler to implement, without the need for jQuery.
Styling and placeholder text have been foregone for simplicity.