fix: add submit button to calendar picker
The calendar picker worked correctly when the calendar popup was used,
however when a date was entered into the text box, the form submitted before the year
could be completed. Also, any date outside the min/max bounds could be entered using
the up/down arrows. Adding this button adds a tiny bit of friction for users, but
with the benefit of input validation.