Update the accounts information page
This adds the account number to the accounts page, which is a critical
piece of information for properly understanding which account is
which.

This version of the page pulls the account number from the configured
role, which is a nice simple way to go about it. A more complete
solution would use the configured role to call `STS.getCallerIdentity`
but this isn't currently required and would make this all more
complicated.

We also tidy up and simplify the "owners" display. The previous
implementation dates back to Janus' early versions, where there were
only "devs" and "admins". This assumption no longer holds, so we
instead list everyone that has any access, and tell users to look at
the Janus access config to find out more detail.

Here again, a future version could show the JSON policy(s) available
to each listed user, but since this duplicates the information in
Janus' configuration there's no need to add this until it is asked
for.