Add suport for managed policies
Janus has historically only supported inline policies, defined within
the Janus codebase. This brings lots of benefits, but limits our
capabilities. Demand for a read-only policy justifies us supporting
managed policies, so this updates the application so that policies
defined in Janus' config can include inline policy definitions and/or
refer to existing policies.

At least initially I'd expect us to limit this AWS managed policies.
This kind of restriction would be a validation of the configuration
rather than the application, and would thus belong in the Janus
config.

We update the Permission type to carry both an inline policy and
managed policies, update the federation code to provide configured
managed policies in assume role requests, and add tests for this new
behaviour.