Allow permissions to specify a unique AWS profile name
By default, Janus bestows temporary credentials into an AWS profile
that is named after the AWS account. This model works well when you're
replacing permanent access to an AWS account with a temporary versino
of the same thing, but is not a good fit for more fine-grained
permissions that cater to a specific use case. This is because you can
only have a single session at a time under one name, so teams that run
multiple concurrent separate highly specific sessions against one
account aren't well-served by Janus today.

It's possible to override the profile name in the UI, but this is a
tedious manual process that isn't sustainable for day-to-day work.

This PR adds a flag to permissions to let Janus know that the profile
name should come from the permission itself, not the AWS account.

The new flag is optional so this would be a backwards compatible
change that adds a new feature for now, but I expect this to become a
more popular way to use Janus in the future.