Let's use a descriptive name for the new permission setting
Instead of using a flag tied to the implementation details of the
profile name override feature, let's think ahead to why we're doing
this and name the property accordingly.

With this in mind, it is now called session type. Creating a session
for an AWS account is the default, and this creates a "user"
session. We also now provide the ability to create permissions that
are tailored to a specific use case, and we expect this to be useful
when running a service locally. These are now called "workload"
sessions.